tools: tests: add a python script to send CAPI commands
Previously, the commands were directly sent to beerocks_cli.  With
wfa_ca moving away from beerocks_cli, there will be a dedicated socket
for the CAPI commands.

This commit adds a python script that makes it easier to send CAPI
commands to the listening socket, from the shell test
script (i.e. test_flows.sh).

The script expects three positional arguments: the host, the port and
the actual command. It connects to the target socket, send the
command, and waits until it receives a COMPLETE, INVALID, or ERROR
reply.
The replies are printed by the script as they are received.
If an INVALID or ERROR reply is received, the script ends with an
exception.
If the device replies with an unexpected string, the script also
ends with an exception.

Signed-off-by: Raphaël Mélotte <raphael.melotte@essensium.com>