tools: docker: set the correct docker IP address
Currently docker containers are assigned an (internal) IP by docker,
but we define and use another one.
When using `docker inspect` to get the IP address of the container,
docker returns the one that it assigned itself (which is not used by
the container).

This commit tells docker which IP address we generated and
assigned ourselves to the container. This allows to later use docker
tools like `docker inspect´ to get the correct IP address.

It does this by splitting $IPADDR into 2 separate IP and netid length
variables, and adding the "--ip" argument when invoking "docker run".

Signed-off-by: Raphaël Mélotte <raphael.melotte@essensium.com>