tools: test_flows: remove extra space from the payload
Previously the payload was built by adding a space then the gateway
mac address to the payload variable.  As the variable was initially
empty, the resulting payload contained an extra space at the
beginning.

This commit initializes the payload variable so that there is no space
at the beginning of the resulting payload. As the payload already
contains one mac address when entering the for loop, it has to run one
time less.

Signed-off-by: Raphaël Mélotte <raphael.melotte@essensium.com>