controller: validate bss info
Currently, test 5.4.3 fails with the following message :

Beacon frame found, check CTT Agent1 is not operating on the configured channel

The UCC tries to validate there are no beacons from that radio, but
beacon frame found.

The reason for this is when the UCC sends the command dev_set_config
without SSID argument it means we should tear down the specified radio,
but the current implementation does not support it.

The solution is to tear down the radio when the UCC sends the
DEV_SET_CONFIG without the SSID parameter, instead of setting the SSID
to an empty string.

Therefore, add a validation check for the bss_info_conf content.

Signed-off-by: Coral Malachi <coral.malachi@intel.com>