common: build with top-level CMakeLists.txt
Add common as a subdirectory in the top-level CMakeLists.txt and update
maptools accordingly.

Since common is now built before framework is installed, we can no
longer rely on find_program to find tlvf.py. Instead, define a custom
propert on the tlvf library to point to the source path of tlvf.py, and
use that property in btlvf.

Since common is now built before framework is installed, we can no
longer use the same find_package for MapfTransport. Change the
conditions to use MSGLIB to determine if we have a transport or not.

Remove the mapf:: prefix for libraries that are now internal instead of
imported.

Because the bcl package depends on tlvf and elpp, the order of
find_package calls becomes important. Therefore, move the tlvf and elpp
find_package calls before the one of bcl in agent and controller.

The cmake helpers in common are no longer needed so remove them.

The top-level external-toolchain files will be used. The only
significant difference was that framework set the default
TARGET_PLATFORM in its CMakeLists.txt, while common set it in
external-toolchain.cfg. Since relying on external-toolchain.cfg is a bad
idea, move the default from framework to top-level and remove it from
external-toolchain.cfg.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>