controller: remove "unknown IP" states
The Turris Omnia doesn't have the IP addresses of the clients, which
make the optimal_path_task fail (it checks for a CONNECTED state, but
the state is CONNECTED_IP_UNKNOWN).

There shouldn't be a difference between a connected state and a
connected state without a known IP, as prplMesh is supposed to work
just as well without IPs.

In cases where it's really needed to know if the device has an IP or
not, it's still posssible to check if the IP has the "0.0.0.0" value.

Note that this commit may need additional fixes, as there are places
where "no IP address" is represented by an empty string (like in the
database), and other places where "no IP address" is represented by
the "0.0.0.0" IP.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>