controller: client steering: use 0 as messageId
The task ID was previously used as the messageId. The consequence was
that the ieee1905_transport was detecting the steering request as
duplicate, because it used the same MAC src, and the same messageId.

This commit uses 0 as messageId, so that it's generated automatically,
and that the packets are no longer flagged as duplicate.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>