encryption: Fix padding by using openssl internal padding
padding was wrong yada yada yada - explain
Update encryption implementation to align to recommended implementation
with default padding done by the library instead of doing so manually.

encryption: use pkcs#7 padding

mapf::encryption::aes_encrypt() implementation disables automatic
padding to be done by openssl API, and instead relies on the user to
make sure that the plaintext is aligned to the underlying aes cbc block
size which is 16 bytes (aes128).

For that, the controller encrypted settings in the WSC M2 TLV includes
zero padding of the config data to 16 bytes boundary.

However, this is not a valid padding and therefore may result with
in decryption failures as seen with some reference golden agents in some
configurations.

TODO - verify that if no padding needed, it always
works

To fix that, we can either replace the encryption implementation to let
the underlying library perform the padding, or select a padding
ourselves.

The most common padding scheme is CMS (Cryptographic Message Syntax).
This pads with the same value as the number of padding bytes. Defined in
RFC 5652, PKCS#5, PKCS#7 (X.509 certificate) and RFC 1423 PEM.

According to [2], "There are multiple methods to pad data, which can
completely undermine the compatibility of a given encryption program
with others. The most used padding method in symmetric ciphers and mode
of operation is certainly PKCS#7".

Therfore, this commit changes the padding from zero padding to PKCS#7.
The result is that encryption is finally correct when tested with online
aes encrption/decryptioni [4].

References:
[1] https://www.di-mgt.com.au/cryptopad.html
[2] https://crypto.stackexchange.com/questions/48628/why-is-padding-used-in-cbc-mode
[3] https://asecuritysite.com/encryption/padding
[4] http://aes.online-domain-tools.com/

Signed-off-by: Tomer Eliyahu <tomer.b.eliyahu@intel.com>

add debug

Signed-off-by: Tomer Eliyahu <tomer.b.eliyahu@intel.com>

Add padding even if plaintext is aligned to block size

Qualcomm agent needs the padding even if the plaintext is aligned.
Verified that this makes Qualcomm work always and doesn't hurt broadcomm
and marvell.
Padding scheme:
if padding needed, use PKCS#7. If padding not needed, still use PKCS#7
padding for exactly block size.

Finally it might be better to let the library do the padding.. Need to
think about it.

Signed-off-by: Tomer Eliyahu <tomer.b.eliyahu@intel.com>

TBD - describe. Fix encryption

Signed-off-by: Tomer Eliyahu <tomer.b.eliyahu@intel.com>