framework: move dump_buffer to mapf::common
dump_buffer is used for dumping raw buffers and is especially useful for
debugging encryption issues, since the format in which it dumps can be
simply copied to online tools which can verify the validity of our
encryption.
Ofcourese, dump_buffer is generic and not bound for being used only in
encryption oriented code, so create a new utils.cpp and utils.h files in
mapf::common library and put it there.
In the future, we will move some functionality (TBD) from BCL to
mapf::common, if not all of it.

Signed-off-by: Tomer Eliyahu <tomer.b.eliyahu@intel.com>