encryption: Use openssl internal padding
Previousely, the aes_encrypt() required the plaintext to be aligned to
block size (16 bytes due to aes128).
Due to this, the config data was zero padded before encryption, which
was wrong since zero padding can only be used if the 0 can't be part of
the plaintext.

The correct padding is described in the WSC specifiction [5]:
"ENCkeywrapkey(...) This notation indicates symmetric encryption of the
values in parentheses using the key keywrapkey. The encryption algorithm
is AES-CBC per FIPS 197, *with PKCS#5 v2.0 padding*".
This padding scheme simply defines each padded character to be the value
of the number of padded characters.

This was the root cause for various failures in autoconfiguration flow
with multiple agents.
Since AES-CBC is a block cipher, it encrypts each block seperately.
So, if the input plaintext padding is wrong, the output ciphertext will
only be wrong in the last block.
Last block checking is optional in openssl (enabled by default), but is
*disabled* in BBF reference implementation, causing most agents to
*accept* invalid encrypted settings and fail spradically since the last
block is all messed up.
Only Qualcomm agent apparently let openssl (or using some internal
impelemntation) perform correctness checks for the last blcoks, which is
why prplMesh never could pass autoconfiguration with Qualcomm agent.

This commit solves all these issues, thus allowing all tests involving
Qualcomm as agent1 or agent2 to pass.
In addition, it solves issues observed with other agents when changing
default SSID which resulted with different padding which caused sporadic
failures in Broadcomm, Marvell, and Mediatek agents.

It does so by letting openssl take care of internal padding, but still
verifies there is enough room in the given ciphertext output buffer for
the necessary padding.

Fixes #192
Fixes #195

References
[1] https://www.di-mgt.com.au/cryptopad.html
[2] https://crypto.stackexchange.com/questions/48628/why-is-padding-used-in-cbc-mode
[3] https://asecuritysite.com/encryption/padding
[4] http://aes.online-domain-tools.com/
[5] WiFi Simple Configuration Technical Specification v2.0.6

Signed-off-by: Tomer Eliyahu <tomer.b.eliyahu@intel.com>
Co-authored-by: Coral Malachi <coral.malachi@intel.com>